+++ src/backend/core/tests/documents/test_api_documents_trashbin.py
@@ -106,6 +106,7 @@ def test_api_documents_trashbin_format():
                     "leave": False,
                     "media_auth": False,
                     "media_check": False,
+                    "mention": False,
                     "move": False,  # Can't move a deleted document
                     "partial_update": False,
                     "restore": True,
@@ -174,6 +175,7 @@ def test_api_documents_trashbin_format():
                     "leave": False,
                     "media_auth": False,
                     "media_check": False,
+                    "mention": False,
                     "move": False,  # Can't move a deleted document
                     "partial_update": False,
                     "restore": True,